+++ src/globals.cpp
@@ -21,3 +21,6 @@ int g_windowPosY = 0;
 
 //Keep streaming data when window is closed
 bool g_sendDataAfterWindowClose = 0;
+
+//Copy NMEA sentence to clipboard when clicking "Send"
+bool g_copyNmeaToClipboard = 0;