+++ src/settings_ui_derived.cpp
@@ -20,6 +20,7 @@ void DialogSettings::OnInit(wxInitDialogEvent& event)
   m_checkBox_restoreWindowPos->SetValue(g_restoreWindowPos);
   m_checkBox_restoreWindowSize->SetValue(g_restoreWindowSize);
   m_checkBox_keepSendingData->SetValue(g_sendDataAfterWindowClose);
+  m_checkBox_copyNmeaToClipboard->SetValue(g_copyNmeaToClipboard);
 
   Layout();
   Fit();
@@ -37,6 +38,7 @@ void DialogSettings::updateSettings()
   g_restoreWindowPos = m_checkBox_restoreWindowPos->IsChecked();
   g_restoreWindowSize = m_checkBox_restoreWindowSize->IsChecked();
   g_sendDataAfterWindowClose = m_checkBox_keepSendingData->IsChecked();
+  g_copyNmeaToClipboard = m_checkBox_copyNmeaToClipboard->IsChecked();
 }